--- src/renderer/components/Experiment/Tasks/QueueTaskModal.tsx
@@ -1,2174 +1,0 @@
-import * as React from 'react';
-import Modal from '@mui/joy/Modal';
-import DialogTitle from '@mui/joy/DialogTitle';
-import DialogContent from '@mui/joy/DialogContent';
-import DialogActions from '@mui/joy/DialogActions';
-import Button from '@mui/joy/Button';
-import FormControl from '@mui/joy/FormControl';
-import FormLabel from '@mui/joy/FormLabel';
-import Input from '@mui/joy/Input';
-import {
-  Alert,
-  FormHelperText,
-  ModalClose,
-  ModalDialog,
-  Stack,
-  Typography,
-  Divider,
-  Slider,
-  Switch,
-  Radio,
-  RadioGroup,
-  Select,
-  Option,
-  Checkbox,
-  Chip,
-  Textarea,
-} from '@mui/joy';
-import Editor from 'renderer/components/Shared/LazyMonacoEditor';
-import {
-  PlayIcon,
-  AlertTriangleIcon,
-  CheckCircleIcon,
-  ChevronDownIcon,
-} from 'lucide-react';
-import { setTheme } from 'renderer/lib/monacoConfig';
-import { useSWRWithAuth as useSWR, useAPI } from 'renderer/lib/authContext';
-import * as chatAPI from 'renderer/lib/transformerlab-api-sdk';
-import { fetcher, getAPIFullPath } from 'renderer/lib/transformerlab-api-sdk';
-import { useAuth } from 'renderer/lib/authContext';
-import SweepConfigSection from './SweepConfigSection';
-import SafeJSONParse from 'renderer/components/Shared/SafeJSONParse';
-import { getPreferredProviderId } from './providerCompatibility';
-
-type QueueTaskModalProps = {
-  open: boolean;
-  onClose: () => void;
-  task: any;
-  onSubmit: (config: Record<string, any>) => void;
-  isSubmitting?: boolean;
-  experimentId?: string;
-};
-
-// Type definitions for parameter schemas
-type ParameterType =
-  | 'int'
-  | 'integer'
-  | 'float'
-  | 'number'
-  | 'bool'
-  | 'boolean'
-  | 'enum'
-  | 'string';
-type UIWidgetType =
-  | 'slider'
-  | 'range'
-  | 'switch'
-  | 'radio'
-  | 'password'
-  | 'select'
-  | 'lab_model_select'
-  | 'lab_dataset_select';
-
-interface ParameterSchema {
-  type?: ParameterType;
-  default?: any;
-  min?: number;
-  max?: number;
-  step?: number;
-  options?: string[];
-  enum?: string[];
-  ui_widget?: UIWidgetType;
-  title?: string;
-  required?: boolean;
-}
-
-interface ProcessedParameter {
-  key: string;
-  value: any;
-  schema: ParameterSchema | null;
-  isShorthand: boolean;
-}
-
-type ProviderResourceGroup = {
-  id: string;
-  name: string;
-  cpus?: string;
-  memory?: string;
-  disk_space?: string;
-  accelerators?: string;
-  num_nodes?: string;
-};
-
-export default function QueueTaskModal({
-  open,
-  onClose,
-  task,
-  onSubmit,
-  isSubmitting = false,
-  experimentId = '',
-}: QueueTaskModalProps) {
-  const { team } = useAuth();
-  const [parameters, setParameters] = React.useState<ProcessedParameter[]>([]);
-  const [customModelDataset, setCustomModelDataset] = React.useState<
-    Set<number>
-  >(new Set());
-  const [validationErrors, setValidationErrors] = React.useState<
-    Record<number, string>
-  >({});
-  const [selectedProviderId, setSelectedProviderId] = React.useState('');
-  const [runSweeps, setRunSweeps] = React.useState(false);
-  const [sweepConfig, setSweepConfig] = React.useState<Record<string, any[]>>(
-    {},
-  );
-  const [sweepMetric, setSweepMetric] = React.useState('eval/loss');
-  const [lowerIsBetter, setLowerIsBetter] = React.useState(true);
-  const [jobSlurmFlags, setJobSlurmFlags] = React.useState<string[]>(['']);
-  const [jobDockerImage, setJobDockerImage] = React.useState('');
-  const [jobRegion, setJobRegion] = React.useState('');
-  const [jobUseSpot, setJobUseSpot] = React.useState(false);
-  const [jobDstackFleetName, setJobDstackFleetName] = React.useState('');
-  const [useTrackio, setUseTrackio] = React.useState(false);
-  const [useProfiling, setUseProfiling] = React.useState(false);
-  const [useProfilingTorch, setUseProfilingTorch] = React.useState(false);
-  const [trackioProjectName, setTrackioProjectName] = React.useState('');
-  const [cpusInput, setCpusInput] = React.useState('');
-  const [memoryInput, setMemoryInput] = React.useState('');
-  const [diskSpaceInput, setDiskSpaceInput] = React.useState('');
-  const [acceleratorsInput, setAcceleratorsInput] = React.useState('');
-  const [numNodesInput, setNumNodesInput] = React.useState('');
-  const [minutesRequestedInput, setMinutesRequestedInput] = React.useState('');
-  const [showResourceOverrides, setShowResourceOverrides] =
-    React.useState(false);
-  const [resourceOverrideMode, setResourceOverrideMode] = React.useState<
-    'manual' | 'group'
-  >('manual');
-  const [selectedResourceGroupId, setSelectedResourceGroupId] =
-    React.useState('');
-  const [resourceGroupCustomized, setResourceGroupCustomized] =
-    React.useState(false);
-  const [showParameterOverrides, setShowParameterOverrides] =
-    React.useState(true);
-  const [jobDescription, setJobDescription] = React.useState('');
-  const [showDescription, setShowDescription] = React.useState(false);
-  const resourceOverridesRef = React.useRef<HTMLDivElement | null>(null);
-  const loadingMessages = React.useMemo(
-    () => [
-      'Contacting compute provider…',
-      'Reserving resources…',
-      'Preparing environment…',
-      'Submitting job configuration…',
-      'Waiting for job ID…',
-    ],
-    [],
-  );
-  const [loadingMessageIndex, setLoadingMessageIndex] = React.useState(0);
-
-  React.useEffect(() => {
-    if (!open || !isSubmitting || loadingMessages.length === 0) {
-      return;
-    }
-
-    // Reset to first message whenever a new submission starts
-    setLoadingMessageIndex(0);
-
-    const interval = window.setInterval(() => {
-      setLoadingMessageIndex((prev) => {
-        const lastIndex = loadingMessages.length - 1;
-        // Once we reach the final message ("Waiting for job ID…"),
-        // stay there instead of looping back to the beginning.
-        if (prev >= lastIndex) {
-          return lastIndex;
-        }
-        return prev + 1;
-      });
-    }, 1500);
-
-    return () => {
-      window.clearInterval(interval);
-    };
-  }, [open, isSubmitting, loadingMessages]);
-
-  React.useEffect(() => {
-    if (!showResourceOverrides) {
-      return;
-    }
-    window.requestAnimationFrame(() => {
-      resourceOverridesRef.current?.scrollIntoView({
-        behavior: 'smooth',
-        block: 'nearest',
-      });
-    });
-  }, [showResourceOverrides]);
-
-  // Fetch available models and datasets from the API
-  const { data: modelsData } = useSWR(
-    open ? chatAPI.Endpoints.Models.LocalList() : null,
-    fetcher,
-  );
-  const { data: datasetsData } = useSWR(
-    open ? chatAPI.Endpoints.Dataset.LocalList() : null,
-    fetcher,
-  );
-
-  // Fetch existing Trackio project names for this experiment (when Trackio enabled)
-  const trackioProjectsKey =
-    open && useTrackio && experimentId
-      ? `${chatAPI.API_URL()}trackio/projects?experiment_id=${encodeURIComponent(experimentId)}`
-      : null;
-  const { data: trackioProjectsData } = useSWR(trackioProjectsKey, fetcher);
-  const trackioProjects: string[] = Array.isArray(trackioProjectsData?.projects)
-    ? trackioProjectsData.projects
-    : [];
-
-  // Fetch available providers
-  const {
-    data: providerListData,
-    error: providerListError,
-    isLoading: providersIsLoading,
-  } = useAPI('compute_provider', ['list'], { teamId: team?.id ?? null });
-
-  const providers = React.useMemo(
-    () => (Array.isArray(providerListData) ? providerListData : []),
-    [providerListData],
-  );
-
-  // Determine if the selected provider is a local provider (must be before isProviderCompatible)
-  const selectedProvider = React.useMemo(
-    () => providers.find((p: any) => p.id === selectedProviderId),
-    [providers, selectedProviderId],
-  );
-  const isLocalProvider = selectedProvider?.type === 'local';
-  const isSlurmProvider = selectedProvider?.type === 'slurm';
-  const isSkypilotProvider = selectedProvider?.type === 'skypilot';
-  const isDstackProvider = selectedProvider?.type === 'dstack';
-  const isGalleryImported = Boolean((task as any)?.gallery_import);
-  const providerResourceGroups = React.useMemo<ProviderResourceGroup[]>(() => {
-    const groups = (selectedProvider?.config as any)?.resource_groups;
-    if (!Array.isArray(groups)) {
-      return [];
-    }
-    return groups
-      .filter((group: any) => group && typeof group === 'object')
-      .map((group: any) => ({
-        id: String(group.id || ''),
-        name: String(group.name || ''),
-        cpus: group.cpus != null ? String(group.cpus) : undefined,
-        memory: group.memory != null ? String(group.memory) : undefined,
-        disk_space:
-          group.disk_space != null ? String(group.disk_space) : undefined,
-        accelerators:
-          group.accelerators != null ? String(group.accelerators) : undefined,
-        num_nodes:
-          group.num_nodes != null ? String(group.num_nodes) : undefined,
-      }))
-      .filter((group: ProviderResourceGroup) => group.id && group.name);
-  }, [selectedProvider]);
-
-  React.useEffect(() => {
-    if (!selectedResourceGroupId) {
-      return;
-    }
-    const stillExists = providerResourceGroups.some(
-      (group) => group.id === selectedResourceGroupId,
-    );
-    if (!stillExists) {
-      setSelectedResourceGroupId('');
-      setResourceOverrideMode('manual');
-      setResourceGroupCustomized(false);
-    }
-  }, [providerResourceGroups, selectedResourceGroupId]);
-
-  const markCustomResourceOverride = React.useCallback(() => {
-    if (resourceOverrideMode === 'group' && selectedResourceGroupId) {
-      setResourceGroupCustomized(true);
-    }
-  }, [resourceOverrideMode, selectedResourceGroupId]);
-
-  const suggestedGalleryResources = React.useMemo(() => {
-    if (!isGalleryImported) return null;
-
-    const mapping = (task as any)?.supportedAccelerators as
-      | Record<
-          string,
-          { resources?: Record<string, unknown> | undefined } | undefined
-        >
-      | undefined;
-    if (!mapping || typeof mapping !== 'object') return null;
-
-    const providerSupported = selectedProvider?.config?.supported_accelerators;
-    if (!Array.isArray(providerSupported)) return null;
-
-    const providerSupportedLower = new Set(
-      providerSupported.map((s: any) => String(s).toLowerCase()),
-    );
-
-    const priority = ['NVIDIA', 'AMD', 'AppleSilicon', 'cpu'];
-    for (const candidate of priority) {
-      const candidateLower = candidate.toLowerCase();
-      if (!providerSupportedLower.has(candidateLower)) continue;
-
-      const entry = (mapping as any)[candidate];
-      const resources = entry?.resources;
-      if (!resources || typeof resources !== 'object') continue;
-
-      const accelerators = (resources as any).accelerators;
-      const cpus = (resources as any).cpus;
-      const memory = (resources as any).memory;
-
-      return {
-        category: candidate,
-        accelerators: accelerators ? String(accelerators) : undefined,
-        cpus: cpus ? String(cpus) : undefined,
-        memory: memory ? String(memory) : undefined,
-      };
-    }
-
-    return null;
-  }, [isGalleryImported, selectedProvider, task]);
-
-  // Fetch user-specific provider settings (including default custom SBATCH flags)
-  const slurmUserSettingsKey =
-    open && isSlurmProvider && selectedProviderId
-      ? getAPIFullPath('compute_provider', ['user-settings'], {
-          providerId: selectedProviderId,
-        })
-      : null;
-  const { data: slurmUserSettings } = useSWR(slurmUserSettingsKey, fetcher);
-
-  // Fetch local provider snapshot via the same response type as remote providers (/compute_provider/{id}/clusters).
-  const clustersKey =
-    open && isLocalProvider && selectedProviderId
-      ? getAPIFullPath('compute_provider', ['providerClusters'], {
-          providerId: selectedProviderId,
-        })
-      : null;
-  const { data: providerClustersData } = useSWR(clustersKey, fetcher);
-  const localProviderConfig = React.useMemo(() => {
-    const clusters = Array.isArray(providerClustersData)
-      ? providerClustersData
-      : [];
-    const localCluster =
-      clusters.find(
-        (c: any) => String(c?.backend_type).toLowerCase() === 'local',
-      ) ??
-      clusters[0] ??
-      null;
-    return (localCluster?.provider_data as any) ?? null;
-  }, [providerClustersData]);
-
-  // Extract task resource requirements from the task object (must be before isProviderCompatible)
-  const taskResources = React.useMemo(() => {
-    if (!task) return null;
-    const cfg =
-      task.config !== undefined ? SafeJSONParse(task.config, task) : task;
-
-    let accelerators = cfg.accelerators || task.accelerators || null;
-    const cpus = cfg.cpus || task.cpus || null;
-    const memory = cfg.memory || task.memory || null;
-
-    // The YAML resources.compute_provider field can contain an accelerator
-    // spec like "RTX3090:1" — the backend stores it as provider_name.
-    // If there's no explicit accelerators field, check if compute_provider /
-    // provider_name looks like an accelerator spec (e.g. "Name:Count").
-    if (!accelerators) {
-      const computeProvider =
-        cfg.compute_provider ||
-        task.compute_provider ||
-        cfg.provider_name ||
-        task.provider_name ||
-        null;
-      if (computeProvider && /^.+:\d+$/.test(String(computeProvider))) {
-        accelerators = String(computeProvider);
-      }
-    }
-
-    if (!accelerators && !cpus && !memory) return null;
-
-    return { accelerators, cpus, memory };
-  }, [task]);
-
-  const effectiveResources = React.useMemo(() => {
-    if (!taskResources) {
-      const acceleratorsEff =
-        acceleratorsInput?.trim() !== '' ? acceleratorsInput.trim() : null;
-      const cpusEff = cpusInput?.trim() !== '' ? cpusInput.trim() : null;
-      const memoryEff = memoryInput?.trim() !== '' ? memoryInput.trim() : null;
-      if (!acceleratorsEff && !cpusEff && !memoryEff) {
-        return null;
-      }
-      return {
-        accelerators: acceleratorsEff,
-        cpus: cpusEff,
-        memory: memoryEff,
-      };
-    }
-
-    const base = taskResources;
-    const cpusEff =
-      cpusInput.trim() !== '' ? cpusInput.trim() : (base.cpus as any);
-    const memoryEff =
-      memoryInput.trim() !== '' ? memoryInput.trim() : (base.memory as any);
-    const acceleratorsEff =
-      acceleratorsInput.trim() !== ''
-        ? acceleratorsInput.trim()
-        : (base.accelerators as any);
-
-    if (!acceleratorsEff && !cpusEff && !memoryEff) {
-      return null;
-    }
-
-    return {
-      accelerators: acceleratorsEff,
-      cpus: cpusEff,
-      memory: memoryEff,
-    };
-  }, [taskResources, cpusInput, memoryInput, acceleratorsInput]);
-
-  // Helper to check if a provider supports requested accelerators
-  const isProviderCompatible = React.useCallback(
-    (provider: any) => {
-      // Only enforce accelerator compatibility heuristics for local providers.
-      // Remote providers (Runpod, Skypilot, SLURM, etc.) validate resources on their side.
-      if (provider?.type !== 'local') {
-        return true;
-      }
-
-      if (!effectiveResources || !effectiveResources.accelerators) return true;
-
-      const supported = provider.config?.supported_accelerators || [];
-      if (supported.length === 0) return true; // Default to compatible if not specified
-
-      const reqAcc = String(effectiveResources.accelerators).toLowerCase();
-
-      // Check for Apple Silicon
-      if (
-        (reqAcc.includes('apple') || reqAcc.includes('mps')) &&
-        supported.includes('AppleSilicon')
-      ) {
-        return true;
-      }
-
-      // Check for NVIDIA
-      if (
-        (reqAcc.includes('nvidia') ||
-          reqAcc.includes('cuda') ||
-          reqAcc.includes('rtx') ||
-          reqAcc.includes('a100') ||
-          reqAcc.includes('h100') ||
-          reqAcc.includes('v100')) &&
-        supported.includes('NVIDIA')
-      ) {
-        return true;
-      }
-
-      // Check for AMD
-      if (
-        (reqAcc.includes('amd') || reqAcc.includes('rocm')) &&
-        supported.includes('AMD')
-      ) {
-        return true;
-      }
-
-      // Check for CPU
-      if (reqAcc.includes('cpu') && supported.includes('cpu')) {
-        return true;
-      }
-
-      // If it's just a number, we assume it's NVIDIA/CUDA unless it's a local provider on Mac
-      if (/^\d+$/.test(reqAcc)) {
-        if (
-          provider.type === 'local' &&
-          localProviderConfig?.device === 'mps'
-        ) {
-          return supported.includes('AppleSilicon');
-        }
-        return supported.includes('NVIDIA');
-      }
-
-      return false;
-    },
-    [effectiveResources, localProviderConfig],
-  );
-
-  React.useEffect(() => {
-    if (providerListError) {
-      // eslint-disable-next-line no-console
-      console.error('Failed to fetch providers', providerListError);
-    }
-  }, [providerListError]);
-
-  const models = modelsData || [];
-  const datasets = datasetsData || [];
-
-  // Validate local provider resources against task requirements
-  const resourceValidation = React.useMemo(() => {
-    if (!isLocalProvider || !localProviderConfig || !effectiveResources)
-      return null;
-
-    const issues: Array<{
-      type: 'error' | 'warning';
-      label: string;
-      required: string;
-      available: string;
-    }> = [];
-
-    // Check accelerators / GPU requirement
-    if (effectiveResources.accelerators) {
-      const accStr = String(effectiveResources.accelerators);
-      // Parse accelerator spec like "RTX3090:1", "A100:2", "V100:4", or just "1" (count only)
-      const match = accStr.match(/^(.+):(\d+)$/);
-      const gpuList: any[] = localProviderConfig.gpu || [];
-      const deviceType = localProviderConfig.device_type || 'cpu';
-
-      if (match) {
-        // Named GPU requirement e.g. "RTX3090:1"
-        const requiredGpuName = match[1].trim();
-        const requiredCount = parseInt(match[2], 10);
-
-        // Count matching GPUs on local machine (fuzzy name match)
-        const matchingGpus = gpuList.filter(
-          (g: any) =>
-            g.name &&
-            g.name !== 'cpu' &&
-            g.name.toLowerCase().includes(requiredGpuName.toLowerCase()),
-        );
-
-        if (
-          deviceType === 'cpu' ||
-          gpuList.length === 0 ||
-          (gpuList.length === 1 && gpuList[0]?.name === 'cpu')
-        ) {
-          issues.push({
-            type: 'error',
-            label: 'GPU',
-            required: `${requiredGpuName} ×${requiredCount}`,
-            available: 'No GPU detected',
-          });
-        } else if (matchingGpus.length === 0) {
-          const availableNames = [
-            ...new Set(
-              gpuList
-                .map((g: any) => g.name)
-                .filter((n: string) => n && n !== 'cpu'),
-            ),
-          ];
-          issues.push({
-            type: 'warning',
-            label: 'GPU',
-            required: `${requiredGpuName} ×${requiredCount}`,
-            available:
-              availableNames.length > 0
-                ? `${availableNames.join(', ')} ×${gpuList.filter((g: any) => g.name !== 'cpu').length}`
-                : 'None',
-          });
-        } else if (matchingGpus.length < requiredCount) {
-          issues.push({
-            type: 'warning',
-            label: 'GPU Count',
-            required: `${requiredGpuName} ×${requiredCount}`,
-            available: `${requiredGpuName} ×${matchingGpus.length}`,
-          });
-        }
-      } else {
-        // Just a count or a name without count
-        const requiredCount = parseInt(accStr, 10);
-        if (!isNaN(requiredCount) && requiredCount > 0) {
-          const realGpus = gpuList.filter(
-            (g: any) => g.name && g.name !== 'cpu',
-          );
-          if (realGpus.length < requiredCount) {
-            issues.push({
-              type: realGpus.length === 0 ? 'error' : 'warning',
-              label: 'GPU',
-              required: `${requiredCount} GPU(s)`,
-              available:
-                realGpus.length === 0
-                  ? 'No GPU detected'
-                  : `${realGpus.length} GPU(s)`,
-            });
-          }
-        }
-      }
-    }
-
-    // Note: We intentionally skip validating CPU and memory for local providers.
-    // Only accelerator (GPU) requirements are enforced here.
-
-    const hasErrors = issues.some((i) => i.type === 'error');
-    const hasWarnings = issues.some((i) => i.type === 'warning');
-
-    return {
-      issues,
-      hasErrors,
-      hasWarnings,
-      isCompatible: issues.length === 0,
-    };
-  }, [isLocalProvider, localProviderConfig, effectiveResources]);
-
-  // Helper function to parse parameter value and schema
-  const parseParameter = (key: string, value: any): ProcessedParameter => {
-    // Check if value is a schema object (has type, default, etc.) or shorthand
-    const isObject =
-      typeof value === 'object' && value !== null && !Array.isArray(value);
-    const hasSchemaFields =
-      isObject &&
-      ('type' in value ||
-        'default' in value ||
-        'min' in value ||
-        'max' in value ||
-        'options' in value ||
-        'enum' in value);
-
-    if (hasSchemaFields) {
-      // Extended schema format
-      const schema = value as ParameterSchema;
-      return {
-        key,
-        value: schema.default !== undefined ? schema.default : '',
-        schema,
-        isShorthand: false,
-      };
-    } else {
-      // Shorthand format - infer type from value
-      return {
-        key,
-        value: value,
-        schema: null,
-        isShorthand: true,
-      };
-    }
-  };
-
-  // Initialize parameters and provider from task when modal opens
-  React.useEffect(() => {
-    if (open && task) {
-      // Extract parameters from task
-      const cfg =
-        task.config !== undefined ? SafeJSONParse(task.config, task) : task;
-
-      const taskParameters = cfg.parameters || task.parameters || {};
-
-      // Convert parameters object to ProcessedParameter array
-      if (typeof taskParameters === 'object' && taskParameters !== null) {
-        const parametersArray = Object.entries(taskParameters).map(
-          ([key, value]) => parseParameter(key, value),
-        );
-
-        setParameters(parametersArray);
-      } else {
-        setParameters([]);
-      }
-
-      // Set provider: use task's provider if it exists, otherwise default provider, then first provider.
-      const taskProviderId = cfg.provider_id ?? task.provider_id ?? '';
-      const taskProviderInList = providers.some(
-        (p: { id: string }) => p.id === taskProviderId,
-      );
-      setSelectedProviderId(
-        taskProviderInList ? taskProviderId : getPreferredProviderId(providers),
-      );
-
-      // Initialize sweep configuration from task
-      setRunSweeps(cfg.run_sweeps ?? task.run_sweeps ?? false);
-      if (cfg.sweep_config || task.sweep_config) {
-        const sweepCfg =
-          typeof cfg.sweep_config === 'string'
-            ? SafeJSONParse(cfg.sweep_config, cfg.sweep_config)
-            : cfg.sweep_config || task.sweep_config;
-        setSweepConfig(sweepCfg || {});
-      } else {
-        setSweepConfig({});
-      }
-      setSweepMetric(
-        cfg.sweep_metric ||
-          task.sweep_metric ||
-          (cfg.run_sweeps || task.run_sweeps ? 'eval/loss' : 'eval/loss'),
-      );
-      setLowerIsBetter(
-        cfg.lower_is_better !== undefined
-          ? cfg.lower_is_better
-          : task.lower_is_better !== undefined
-            ? task.lower_is_better
-            : true,
-      );
-      const initCpus = cfg.cpus ?? task.cpus ?? '';
-      const initMemory = cfg.memory ?? task.memory ?? '';
-      const initDiskSpace = cfg.disk_space ?? task.disk_space ?? '';
-      const initAccelerators = cfg.accelerators ?? task.accelerators ?? '';
-      const initNumNodes = cfg.num_nodes ?? task.num_nodes ?? '';
-      const initMinutesRequested =
-        cfg.minutes_requested ?? task.minutes_requested ?? '';
-
-      setCpusInput(
-        initCpus !== null && initCpus !== undefined ? String(initCpus) : '',
-      );
-      setMemoryInput(
-        initMemory !== null && initMemory !== undefined
-          ? String(initMemory)
-          : '',
-      );
-      setDiskSpaceInput(
-        initDiskSpace !== null && initDiskSpace !== undefined
-          ? String(initDiskSpace)
-          : '',
-      );
-      setAcceleratorsInput(
-        initAccelerators !== null && initAccelerators !== undefined
-          ? String(initAccelerators)
-          : '',
-      );
-      setNumNodesInput(
-        initNumNodes !== null && initNumNodes !== undefined
-          ? String(initNumNodes)
-          : '',
-      );
-      setMinutesRequestedInput(
-        initMinutesRequested !== null && initMinutesRequested !== undefined
-          ? String(initMinutesRequested)
-          : '',
-      );
-      setResourceOverrideMode('manual');
-      setSelectedResourceGroupId('');
-      setResourceGroupCustomized(false);
-      const initDescription = cfg.description ?? task.description ?? '';
-      const normalizedDescription =
-        initDescription !== null && initDescription !== undefined
-          ? String(initDescription)
-          : '';
-      setJobDescription(normalizedDescription);
-      setShowDescription(Boolean(normalizedDescription.trim()));
-    }
-  }, [open, task, providers]);
-
-  // When opening for a SLURM provider, initialize job flags from the user's
-  // saved per-provider defaults so they can be edited for this run.
-  React.useEffect(() => {
-    if (!open || !isSlurmProvider || !slurmUserSettings) return;
-    const raw = (slurmUserSettings as any).custom_sbatch_flags || '';
-    const lines = String(raw)
-      .split('\n')
-      .map((line) => line.trim())
-      .filter((line) => line.length > 0);
-    setJobSlurmFlags(lines.length > 0 ? lines : ['']);
-  }, [open, isSlurmProvider, selectedProviderId, slurmUserSettings]);
-
-  // Initialize SkyPilot per-job defaults from provider config when a SkyPilot provider is selected.
-  React.useEffect(() => {
-    if (!open || !isSkypilotProvider || !selectedProvider) return;
-    const cfg = selectedProvider.config || {};
-    setJobDockerImage(cfg.docker_image || '');
-    setJobRegion(cfg.default_region || '');
-    setJobUseSpot(cfg.use_spot === true);
-  }, [open, isSkypilotProvider, selectedProviderId, selectedProvider]);
-
-  // Initialize dstack per-job defaults from provider/task config when selected.
-  React.useEffect(() => {
-    if (!open || !isDstackProvider || !selectedProvider) return;
-    const providerCfg = selectedProvider.config || {};
-    const taskCfg =
-      task?.config !== undefined ? SafeJSONParse(task.config, {}) : {};
-    const taskRunCfg = taskCfg?.config || {};
-    const taskResourcesCfg = taskRunCfg?.resources || {};
-    setJobDstackFleetName(
-      String(
-        taskRunCfg.fleet_name ||
-          taskResourcesCfg.fleet_name ||
-          providerCfg.fleet_name ||
-          '',
-      ),
-    );
-  }, [open, isDstackProvider, selectedProviderId, selectedProvider, task]);
-
-  // Helper function to validate constraints
-  const validateParameter = (param: ProcessedParameter): string | null => {
-    const { schema, value } = param;
-    if (!schema) return null;
-
-    const numValue = Number(value);
-
-    // Check min constraint
-    if (schema.min !== undefined && !Number.isNaN(numValue)) {
-      if (numValue < schema.min) {
-        return `Minimum value is ${schema.min}`;
-      }
-    }
-
-    // Check max constraint
-    if (schema.max !== undefined && !Number.isNaN(numValue)) {
-      if (numValue > schema.max) {
-        return `Maximum value is ${schema.max}`;
-      }
-    }
-
-    // Note: step validation is handled by the native HTML input step attribute
-
-    return null;
-  };
-
-  // Helper function to check if all parameters are valid
-  const validateAllParameters = (): string | null => {
-    for (const param of parameters) {
-      if (!param.key.trim()) continue;
-      const error = validateParameter(param);
-      if (error) {
-        return `${param.key}: ${error}`;
-      }
-    }
-    return null;
-  };
-
-  const handleSubmit = () => {
-    // Validate provider selection
-    if (!selectedProviderId) {
-      alert('Please select a compute provider before submitting');
-      return;
-    }
-
-    // Validate all parameters
-    const validationError = validateAllParameters();
-    if (validationError) {
-      alert(`Validation error: ${validationError}`);
-      return;
-    }
-
-    // Validate sweep configuration if enabled
-    if (runSweeps && Object.keys(sweepConfig).length === 0) {
-      alert(
-        'Please add at least one parameter to sweep, or disable hyperparameter sweeps.',
-      );
-      return;
-    }
-
-    // Convert parameters array to object for config
-    const config: Record<string, any> = {};
-    parameters.forEach(({ key, value }) => {
-      if (key.trim()) {
-        config[key.trim()] = value;
-      }
-    });
-
-    // Add provider_id to the config
-    config.provider_id = selectedProviderId;
-
-    // Add provider_name if available
-    const provider = providers.find((p) => p.id === selectedProviderId);
-    const selectedResourceGroup = providerResourceGroups.find(
-      (group) => group.id === selectedResourceGroupId,
-    );
-    if (provider) {
-      config.provider_name = provider.name;
-    }
-
-    if (resourceOverrideMode === 'group' && selectedResourceGroup) {
-      config.resource_group_id = selectedResourceGroup.id;
-      config.resource_group_name = selectedResourceGroup.name;
-      if (resourceGroupCustomized) {
-        config.resource_group_custom_override = true;
-      }
-    }
-
-    if (cpusInput.trim()) {
-      config.cpus = cpusInput.trim();
-    }
-    if (memoryInput.trim()) {
-      config.memory = memoryInput.trim();
-    }
-    if (diskSpaceInput.trim()) {
-      config.disk_space = diskSpaceInput.trim();
-    }
-    if (acceleratorsInput.trim()) {
-      config.accelerators = acceleratorsInput.trim();
-    }
-    if (numNodesInput.trim()) {
-      const parsedNumNodes = Number(numNodesInput.trim());
-      config.num_nodes = Number.isNaN(parsedNumNodes)
-        ? numNodesInput.trim()
-        : parsedNumNodes;
-    }
-    if (minutesRequestedInput.trim()) {
-      const parsedMinutes = Number(minutesRequestedInput.trim());
-      if (!Number.isNaN(parsedMinutes) && parsedMinutes > 0) {
-        config.minutes_requested = parsedMinutes;
-      }
-    }
-    if (jobDescription.trim()) {
-      config.description = jobDescription.trim();
-    }
-
-    // For SLURM providers, add optional per-job SBATCH flags override
-    if (provider?.type === 'slurm') {
-      const lines = jobSlurmFlags
-        .map((line) => line.trim())
-        .filter((line) => line.length > 0);
-      if (lines.length > 0) {
-        config.custom_sbatch_flags = lines.join('\n');
-      }
-    }
-
-    // For SkyPilot providers, add optional per-job overrides
-    if (provider?.type === 'skypilot') {
-      if (jobDockerImage.trim()) {
-        config.docker_image = jobDockerImage.trim();
-      }
-      if (jobRegion.trim()) {
-        config.region = jobRegion.trim();
-      }
-      if (jobUseSpot) {
-        config.use_spot = true;
-      }
-    }
-
-    // For dstack providers, allow per-run fleet override.
-    if (provider?.type === 'dstack' && jobDstackFleetName.trim()) {
-      config.fleet_name = jobDstackFleetName.trim();
-    }
-
-    // Add sweep configuration if enabled
-    if (runSweeps) {
-      config.run_sweeps = true;
-      if (Object.keys(sweepConfig).length > 0) {
-        config.sweep_config = sweepConfig;
-      }
-      config.sweep_metric = sweepMetric;
-      config.lower_is_better = lowerIsBetter;
-    }
-
-    // Trackio auto-init flag: when enabled, backend will set TLAB_TRACKIO_AUTO_INIT
-    // in the job environment so Lab can automatically integrate with Trackio.
-    if (useTrackio) {
-      config.enable_trackio = true;
-      config.trackio_project_name = trackioProjectName.trim() || undefined;
-    }
-
-    // Profiling: when enabled, backend will set _TFL_PROFILING=1 so tfl-remote-trap
-    // samples CPU/GPU/memory during the job and writes profiling_report.json.
-    if (useProfiling) {
-      config.enable_profiling = true;
-      if (useProfilingTorch) {
-        config.enable_profiling_torch = true;
-      }
-    }
-
-    onSubmit(config);
-  };
-
-  const getTaskTitle = () => {
-    if (task?.title && task.title.trim() !== '') {
-      return task.title;
-    }
-    return task?.name || 'Task';
-  };
-
-  // Helper function to determine the actual type of a parameter
-  const getParameterType = (param: ProcessedParameter): ParameterType => {
-    if (param.schema?.type) {
-      return param.schema.type;
-    }
-    // Infer type from value for shorthand
-    const val = param.value;
-    if (typeof val === 'boolean') return 'bool';
-    if (typeof val === 'number') {
-      return Number.isInteger(val) ? 'int' : 'float';
-    }
-    if (Array.isArray(val)) return 'string'; // Will be rendered as JSON editor
-    if (typeof val === 'object' && val !== null) return 'string'; // Will be rendered as JSON editor
-    return 'string';
-  };
-
-  // Helper function to render the appropriate input widget
-  const updateParameterAndValidate = (
-    newParams: ProcessedParameter[],
-    index: number,
-  ) => {
-    setParameters(newParams);
-
-    // Validate the updated parameter
-    const error = validateParameter(newParams[index]);
-    const newErrors = { ...validationErrors };
-    if (error) {
-      newErrors[index] = error;
-    } else {
-      delete newErrors[index];
-    }
-    setValidationErrors(newErrors);
-  };
-
-  // Helper function to render the appropriate input widget
-  const renderParameterInput = (param: ProcessedParameter, index: number) => {
-    const schema = param.schema;
-    const type = getParameterType(param);
-    const uiWidget = schema?.ui_widget;
-    const label = schema?.title || param.key;
-
-    // Special handling for 'lab_model_select' widget
-    if (uiWidget === 'lab_model_select') {
-      const isCustom = customModelDataset.has(index);
-
-      return (
-        <Stack direction="column" spacing={1} sx={{ flex: 1 }}>
-          {isCustom ? (
-            <Input
-              placeholder="Enter any model name"
-              value={String(param.value)}
-              onChange={(e) => {
-                const newParams = [...parameters];
-                newParams[index].value = e.target.value;
-                setParameters(newParams);
-              }}
-              sx={{ flex: 1 }}
-              disabled={isSubmitting}
-            />
-          ) : (
-            <Select
-              value={String(param.value)}
-              onChange={(_, value) => {
-                const newParams = [...parameters];
-                newParams[index].value = value;
-                setParameters(newParams);
-              }}
-              placeholder="Select a model"
-              sx={{ flex: 1 }}
-              disabled={isSubmitting}
-            >
-              {models.map((model: any) => (
-                <Option key={model.model_id} value={model.model_id}>
-                  {model.model_id}
-                </Option>
-              ))}
-            </Select>
-          )}
-          <Stack direction="row" spacing={1} alignItems="center">
-            <Checkbox
-              checked={isCustom}
-              onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
-                if (e.target.checked) {
-                  const newParams = [...parameters];
-                  // Initialize with empty string if value is null/undefined
-                  if (
-                    newParams[index].value === null ||
-                    newParams[index].value === undefined
-                  ) {
-                    newParams[index].value = '';
-                  }
-                  setParameters(newParams);
-                  setCustomModelDataset(
-                    new Set([...customModelDataset, index]),
-                  );
-                } else {
-                  const newSet = new Set(customModelDataset);
-                  newSet.delete(index);
-                  setCustomModelDataset(newSet);
-                }
-              }}
-              size="sm"
-              disabled={isSubmitting}
-            />
-            <Typography level="body-xs" sx={{ color: 'neutral.500' }}>
-              Enter any string
-            </Typography>
-          </Stack>
-        </Stack>
-      );
-    }
-
-    // Special handling for 'lab_dataset_select' widget
-    if (uiWidget === 'lab_dataset_select') {
-      const isCustom = customModelDataset.has(index);
-
-      return (
-        <Stack direction="column" spacing={1} sx={{ flex: 1 }}>
-          {isCustom ? (
-            <Input
-              placeholder="Enter any dataset name"
-              value={String(param.value)}
-              onChange={(e) => {
-                const newParams = [...parameters];
-                newParams[index].value = e.target.value;
-                setParameters(newParams);
-              }}
-              sx={{ flex: 1 }}
-              disabled={isSubmitting}
-            />
-          ) : (
-            <Select
-              value={String(param.value)}
-              onChange={(_, value) => {
-                const newParams = [...parameters];
-                newParams[index].value = value;
-                setParameters(newParams);
-              }}
-              placeholder="Select a dataset"
-              sx={{ flex: 1 }}
-              disabled={isSubmitting}
-            >
-              {datasets.map((dataset: any) => (
-                <Option key={dataset.dataset_id} value={dataset.dataset_id}>
-                  {dataset.dataset_id}
-                </Option>
-              ))}
-            </Select>
-          )}
-          <Stack direction="row" spacing={1} alignItems="center">
-            <Checkbox
-              checked={isCustom}
-              onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
-                if (e.target.checked) {
-                  const newParams = [...parameters];
-                  // Initialize with empty string if value is null/undefined
-                  if (
-                    newParams[index].value === null ||
-                    newParams[index].value === undefined
-                  ) {
-                    newParams[index].value = '';
-                  }
-                  setParameters(newParams);
-                  setCustomModelDataset(
-                    new Set([...customModelDataset, index]),
-                  );
-                } else {
-                  const newSet = new Set(customModelDataset);
-                  newSet.delete(index);
-                  setCustomModelDataset(newSet);
-                }
-              }}
-              size="sm"
-              disabled={isSubmitting}
-            />
-            <Typography level="body-xs" sx={{ color: 'neutral.500' }}>
-              Enter any string
-            </Typography>
-          </Stack>
-        </Stack>
-      );
-    }
-
-    // Handle different widget types
-    // Integer with slider
-    if (
-      (type === 'int' ||
-        type === 'integer' ||
-        type === 'float' ||
-        type === 'number') &&
-      (uiWidget === 'slider' || uiWidget === 'range')
-    ) {
-      const min = schema?.min ?? 0;
-      const max = schema?.max ?? 100;
-      const step =
-        schema?.step ?? (type === 'int' || type === 'integer' ? 1 : 0.01);
-
-      return (
-        <Stack direction="column" spacing={1} sx={{ flex: 1 }}>
-          <Stack direction="row" spacing={2} alignItems="center">
-            <Slider
-              value={Number(param.value) || min}
-              onChange={(_, value) => {
-                const newParams = [...parameters];
-                newParams[index].value = value;
-                updateParameterAndValidate(newParams, index);
-              }}
-              min={min}
-              max={max}
-              step={step}
-              valueLabelDisplay="auto"
-              sx={{ flex: 1 }}
-              disabled={isSubmitting}
-            />
-            <Input
-              value={param.value}
-              onChange={(e) => {
-                const newParams = [...parameters];
-                const val = e.target.value;
-                newParams[index].value = val === '' ? min : Number(val);
-                updateParameterAndValidate(newParams, index);
-              }}
-              type="number"
-              slotProps={{
-                input: {
-                  min,
-                  max,
-                  step,
-                },
-              }}
-              sx={{ width: 100 }}
-              error={!!validationErrors[index]}
-              disabled={isSubmitting}
-            />
-          </Stack>
-          {validationErrors[index] && (
-            <FormHelperText sx={{ color: 'danger.400' }}>
-              {validationErrors[index]}
-            </FormHelperText>
-          )}
-        </Stack>
-      );
-    }
-
-    // Boolean with switch
-    if (
-      (type === 'bool' || type === 'boolean') &&
-      (uiWidget === 'switch' || !uiWidget)
-    ) {
-      return (
-        <Switch
-          checked={Boolean(param.value)}
-          onChange={(e) => {
-            const newParams = [...parameters];
-            newParams[index].value = e.target.checked;
-            setParameters(newParams);
-          }}
-          disabled={isSubmitting}
-        />
-      );
-    }
-
-    // Enum with radio or select
-    if (type === 'enum' || schema?.options || schema?.enum) {
-      const options = schema?.options || schema?.enum || [];
-
-      if (uiWidget === 'radio') {
-        return (
-          <RadioGroup
-            value={String(param.value)}
-            onChange={(e) => {
-              const newParams = [...parameters];
-              newParams[index].value = e.target.value;
-              setParameters(newParams);
-            }}
-          >
-            <Stack direction="row" spacing={2}>
-              {options.map((option) => (
-                <Radio
-                  key={option}
-                  value={option}
-                  label={option}
-                  disabled={isSubmitting}
-                />
-              ))}
-            </Stack>
-          </RadioGroup>
-        );
-      } else {
-        // Default to select dropdown
-        return (
-          <Select
-            value={String(param.value)}
-            onChange={(_, value) => {
-              const newParams = [...parameters];
-              newParams[index].value = value;
-              setParameters(newParams);
-            }}
-            sx={{ flex: 1 }}
-            disabled={isSubmitting}
-          >
-            {options.map((option) => (
-              <Option key={option} value={option}>
-                {option}
-              </Option>
-            ))}
-          </Select>
-        );
-      }
-    }
-
-    // Integer or Float without slider
-    if (
-      type === 'int' ||
-      type === 'integer' ||
-      type === 'float' ||
-      type === 'number'
-    ) {
-      const min = schema?.min;
-      const max = schema?.max;
-      const step =
-        schema?.step ?? (type === 'int' || type === 'integer' ? 1 : 0.00001);
-
-      return (
-        <Stack direction="column" spacing={1} sx={{ flex: 1 }}>
-          <Input
-            type="number"
-            value={param.value}
-            onChange={(e) => {
-              const newParams = [...parameters];
-              const val = e.target.value;
-              newParams[index].value = val === '' ? '' : Number(val);
-              updateParameterAndValidate(newParams, index);
-            }}
-            slotProps={{
-              input: {
-                min,
-                max,
-                step,
-              },
-            }}
-            sx={{ flex: 1 }}
-            error={!!validationErrors[index]}
-            disabled={isSubmitting}
-          />
-          {validationErrors[index] && (
-            <FormHelperText sx={{ color: 'danger.400' }}>
-              {validationErrors[index]}
-            </FormHelperText>
-          )}
-        </Stack>
-      );
-    }
-
-    // String with password widget
-    if (type === 'string' && uiWidget === 'password') {
-      return (
-        <Input
-          type="password"
-          value={String(param.value)}
-          onChange={(e) => {
-            const newParams = [...parameters];
-            newParams[index].value = e.target.value;
-            setParameters(newParams);
-          }}
-          sx={{ flex: 1 }}
-          disabled={isSubmitting}
-        />
-      );
-    }
-
-    // Complex objects/arrays - render as JSON editor
-    if (
-      !param.isShorthand &&
-      (Array.isArray(param.value) ||
-        (typeof param.value === 'object' && param.value !== null))
-    ) {
-      return (
-        <Editor
-          height="120px"
-          defaultLanguage="json"
-          value={JSON.stringify(param.value, null, 2)}
-          onChange={(value) => {
-            const newParams = [...parameters];
-            try {
-              newParams[index].value = value ? JSON.parse(value) : null;
-            } catch (e) {
-              // Keep the string value if parsing fails
-            }
-            setParameters(newParams);
-          }}
-          theme="my-theme"
-          onMount={setTheme}
-          options={{
-            minimap: { enabled: false },
-            scrollBeyondLastLine: false,
-            fontSize: 12,
-            lineNumbers: 'off',
-            wordWrap: 'on',
-            readOnly: isSubmitting,
-          }}
-        />
-      );
-    }
-
-    // Default: regular string input
-    return (
-      <Input
-        placeholder="Value (e.g., 0.001, true, false, or any string)"
-        value={String(param.value)}
-        onChange={(e) => {
-          const newParams = [...parameters];
-          newParams[index].value = e.target.value;
-          setParameters(newParams);
-        }}
-        sx={{ flex: 1 }}
-        disabled={isSubmitting}
-      />
-    );
-  };
-
-  return (
-    <Modal open={open} onClose={onClose}>
-      <ModalDialog
-        sx={{
-          maxWidth: 700,
-          width: '90%',
-          maxHeight: '90vh',
-          overflow: 'auto',
-        }}
-      >
-        <ModalClose />
-        <DialogTitle>Queue Task: {getTaskTitle()}</DialogTitle>
-        <Divider />
-        <DialogContent>
-          <Stack spacing={3}>
-            {/* Run Settings Section */}
-            <Stack spacing={2}>
-              <Typography level="title-sm">Run Settings</Typography>
-              <FormControl required>
-                <FormLabel>Compute Provider</FormLabel>
-                <Select
-                  placeholder={
-                    providers.length
-                      ? 'Select a compute provider'
-                      : 'No compute providers configured'
-                  }
-                  value={selectedProviderId || null}
-                  onChange={(_, value) => {
-                    setSelectedProviderId(value || '');
-                    setSelectedResourceGroupId('');
-                    setResourceOverrideMode('manual');
-                    setResourceGroupCustomized(false);
-                  }}
-                  disabled={
-                    isSubmitting || providersIsLoading || providers.length === 0
-                  }
-                  renderValue={(selected) => {
-                    const value = selected?.value;
-                    const provider = providers.find(
-                      (p: { id: string }) => p.id === value,
-                    );
-                    return (
-                      provider?.name ||
-                      (providers.length
-                        ? 'Select a compute provider'
-                        : 'No compute providers configured')
-                    );
-                  }}
-                  slotProps={{
-                    listbox: { sx: { maxHeight: 240 } },
-                  }}
-                >
-                  {providers.map((provider: any) => (
-                    <Option key={provider.id} value={provider.id}>
-                      <Stack
-                        direction="row"
-                        justifyContent="space-between"
-                        alignItems="center"
-                        sx={{ width: '100%' }}
-                      >
-                        <Typography>{provider.name}</Typography>
-                      </Stack>
-                    </Option>
-                  ))}
-                </Select>
-                <FormHelperText>
-                  Choose which compute provider should run this task.
-                </FormHelperText>
-              </FormControl>
-
-              {isGalleryImported && (
-                <Alert
-                  variant="soft"
-                  color="warning"
-                  startDecorator={<AlertTriangleIcon size={18} />}
-                  sx={{ mt: 1 }}
-                >
-                  <Typography level="body-sm">
-                    This task was imported from the gallery. Please make sure
-                    the selected resources for this task match the
-                    provider&apos;s availability.
-                  </Typography>
-                  {suggestedGalleryResources && (
-                    <Typography level="body-xs" sx={{ mt: 1 }}>
-                      Suggested ({suggestedGalleryResources.category}):{' '}
-                      {[
-                        suggestedGalleryResources.accelerators &&
-                          `accelerators=${suggestedGalleryResources.accelerators}`,
-                        suggestedGalleryResources.cpus &&
-                          `cpus=${suggestedGalleryResources.cpus}`,
-                        suggestedGalleryResources.memory &&
-                          `memory=${suggestedGalleryResources.memory}`,
-                      ]
-                        .filter(Boolean)
-                        .join(', ')}
-                    </Typography>
-                  )}
-                  <Stack direction="row" spacing={1} sx={{ mt: 1 }}>
-                    <Button
-                      size="sm"
-                      variant="plain"
-                      color="warning"
-                      onClick={() => {
-                        if (!showResourceOverrides) {
-                          setShowResourceOverrides(true);
-                          return;
-                        }
-                        resourceOverridesRef.current?.scrollIntoView({
-                          behavior: 'smooth',
-                          block: 'nearest',
-                        });
-                      }}
-                      disabled={isSubmitting}
-                    >
-                      Review resources
-                    </Button>
-                  </Stack>
-                </Alert>
-              )}
-
-              {/* SLURM per-job SBATCH flags */}
-              {isSlurmProvider && (
-                <FormControl>
-                  <FormLabel>Job-specific SBATCH flags (optional)</FormLabel>
-                  <Stack gap={1}>
-                    {jobSlurmFlags.map((value, idx) => (
-                      <Stack
-                        key={idx}
-                        direction="row"
-                        alignItems="center"
-                        gap={1}
-                      >
-                        <Input
-                          placeholder={
-                            idx === 0 ? '--time=4:00:00' : '--ntasks-per-node=4'
-                          }
-                          sx={{ fontFamily: 'monospace', fontSize: 'sm' }}
-                          value={value}
-                          onChange={(e) => {
-                            const next = [...jobSlurmFlags];
-                            next[idx] = e.target.value;
-                            setJobSlurmFlags(next);
-                          }}
-                          disabled={isSubmitting}
-                        />
-                        {jobSlurmFlags.length > 1 && (
-                          <Button
-                            size="sm"
-                            variant="outlined"
-                            color="neutral"
-                            onClick={() => {
-                              const next = jobSlurmFlags.filter(
-                                (_, i) => i !== idx,
-                              );
-                              setJobSlurmFlags(next.length > 0 ? next : ['']);
-                            }}
-                            disabled={isSubmitting}
-                          >
-                            Remove
-                          </Button>
-                        )}
-                      </Stack>
-                    ))}
-                    <Button
-                      size="sm"
-                      variant="outlined"
-                      onClick={() => setJobSlurmFlags((prev) => [...prev, ''])}
-                      disabled={isSubmitting}
-                    >
-                      Add flag for this job
-                    </Button>
-                  </Stack>
-                  <FormHelperText>
-                    These flags apply only to this queued run and are added as
-                    #SBATCH directives in the SLURM script. They start from your
-                    defaults in User Settings → Provider Settings, but edits
-                    here affect this run only. Examples: --time=4:00:00,
-                    --ntasks-per-node=4.
-                  </FormHelperText>
-                </FormControl>
-              )}
-            </Stack>
-
-            <Divider />
-
-            {/* Task Parameters Section */}
-            <Stack spacing={2}>
-              <Stack
-                direction="row"
-                alignItems="center"
-                justifyContent="space-between"
-                sx={{ cursor: isSubmitting ? 'default' : 'pointer' }}
-                onClick={() => {
-                  if (!isSubmitting) {
-                    setShowParameterOverrides((prev) => !prev);
-                  }
-                }}
-              >
-                <Typography level="title-sm">Parameter overrides</Typography>
-                <ChevronDownIcon
-                  size={18}
-                  style={{
-                    transform: showParameterOverrides
-                      ? 'rotate(180deg)'
-                      : 'rotate(0deg)',
-                    transition: 'transform 0.2s ease',
-                  }}
-                />
-              </Stack>
-              {showParameterOverrides && (
-                <Stack spacing={2}>
-                  {parameters.length === 0 ||
-                  (parameters.length === 1 &&
-                    !parameters[0].key &&
-                    !parameters[0].value) ? (
-                    <Typography level="body-sm" color="neutral">
-                      This task has no parameters defined. Click Submit to queue
-                      with default configuration.
-                    </Typography>
-                  ) : (
-                    <Stack spacing={2}>
-                      {parameters.map((param, index) => {
-                        const schema = param.schema;
-                        const label = schema?.title || param.key;
-
-                        return (
-                          <FormControl
-                            key={param.key || index}
-                            sx={{ width: '100%' }}
-                          >
-                            <Stack
-                              direction="row"
-                              spacing={2}
-                              alignItems="center"
-                              sx={{ width: '100%' }}
-                            >
-                              <FormLabel
-                                sx={{ alignSelf: 'center', minWidth: 160 }}
-                              >
-                                {label}:
-                              </FormLabel>
-                              {renderParameterInput(param, index)}
-                            </Stack>
-                          </FormControl>
-                        );
-                      })}
-                    </Stack>
-                  )}
-                  <Typography level="body-sm" color="neutral">
-                    Parameters can be accessed in your task script using{' '}
-                    <code>lab.get_config()</code>
-                  </Typography>
-                </Stack>
-              )}
-            </Stack>
-
-            <Divider />
-
-            {/* Tracking Section */}
-            <Stack spacing={2}>
-              <Typography level="title-sm">Tracking</Typography>
-              <FormControl
-                orientation="horizontal"
-                sx={{ alignItems: 'center' }}
-              >
-                <Checkbox
-                  checked={useTrackio}
-                  onChange={(e) => setUseTrackio(e.target.checked)}
-                  disabled={isSubmitting}
-                />
-                <FormLabel sx={{ ml: 1 }}>
-                  Enable Trackio metrics tracking for this run
-                </FormLabel>
-              </FormControl>
-              {useTrackio && (
-                <FormControl>
-                  <FormLabel>Project name</FormLabel>
-                  <Input
-                    placeholder="e.g. my-finetune-project"
-                    value={trackioProjectName}
-                    onChange={(e) => setTrackioProjectName(e.target.value)}
-                    disabled={isSubmitting}
-                    slotProps={{
-                      input: {
-                        list: 'trackio-projects-list',
-                      },
-                    }}
-                  />
-                  <datalist id="trackio-projects-list">
-                    {trackioProjects.map((p) => (
-                      <option key={p} value={p} />
-                    ))}
-                  </datalist>
-                  <FormHelperText>
-                    Pick an existing project to add this run to it, or type a
-                    new name to create one.
-                  </FormHelperText>
-                </FormControl>
-              )}
-              <FormHelperText>
-                When enabled, the scripts that use the lab SDK can automatically
-                log metrics to Trackio and expose a Trackio dashboard in the UI.
-              </FormHelperText>
-            </Stack>
-
-            <Divider />
-
-            {/* Profiling Section */}
-            <Stack spacing={2}>
-              <Typography level="title-sm">Profiling</Typography>
-              <FormControl
-                orientation="horizontal"
-                sx={{ alignItems: 'center' }}
-              >
-                <Checkbox
-                  checked={useProfiling}
-                  onChange={(e) => {
-                    setUseProfiling(e.target.checked);
-                    if (!e.target.checked) setUseProfilingTorch(false);
-                  }}
-                  disabled={isSubmitting}
-                />
-                <FormLabel sx={{ ml: 1 }}>
-                  Enable CPU &amp; GPU profiling for this run
-                </FormLabel>
-              </FormControl>
-              <FormHelperText>
-                Samples CPU%, memory, and GPU utilization every few seconds
-                during the job. Results are available in the Profiling tab after
-                the job completes.
-              </FormHelperText>
-              {useProfiling && (
-                <FormControl
-                  orientation="horizontal"
-                  sx={{ alignItems: 'center', ml: 3 }}
-                >
-                  <Checkbox
-                    checked={useProfilingTorch}
-                    onChange={(e) => setUseProfilingTorch(e.target.checked)}
-                    disabled={isSubmitting}
-                  />
-                  <FormLabel sx={{ ml: 1 }}>
-                    Also capture PyTorch op-level trace (Chrome trace format)
-                  </FormLabel>
-                </FormControl>
-              )}
-            </Stack>
-
-            <Divider />
-
-            {/* Sweep Configuration Section */}
-            <SweepConfigSection
-              runSweeps={runSweeps}
-              onRunSweepsChange={setRunSweeps}
-              sweepConfig={sweepConfig}
-              onSweepConfigChange={setSweepConfig}
-              sweepMetric={sweepMetric}
-              onSweepMetricChange={setSweepMetric}
-              lowerIsBetter={lowerIsBetter}
-              onLowerIsBetterChange={setLowerIsBetter}
-              parameters={parameters}
-              disabled={isSubmitting}
-            />
-
-            <Divider />
-
-            {/* Optional Description Section */}
-            <Stack spacing={1}>
-              <Stack
-                direction="row"
-                alignItems="center"
-                justifyContent="space-between"
-                sx={{ cursor: isSubmitting ? 'default' : 'pointer' }}
-                onClick={() => {
-                  if (!isSubmitting) {
-                    setShowDescription((prev) => !prev);
-                  }
-                }}
-              >
-                <Typography level="title-sm">
-                  Add description (optional)
-                </Typography>
-                <ChevronDownIcon
-                  size={18}
-                  style={{
-                    transform: showDescription
-                      ? 'rotate(180deg)'
-                      : 'rotate(0deg)',
-                    transition: 'transform 0.2s ease',
-                  }}
-                />
-              </Stack>
-              {showDescription && (
-                <FormControl>
-                  <FormLabel>Run description</FormLabel>
-                  <Textarea
-                    minRows={3}
-                    maxRows={8}
-                    placeholder="Describe what changed, your hypothesis, and what to watch for in this run."
-                    value={jobDescription}
-                    onChange={(e) => setJobDescription(e.target.value)}
-                    disabled={isSubmitting}
-                  />
-                  <FormHelperText>
-                    Stored with the job and shown in job details. Markdown is
-                    supported.
-                  </FormHelperText>
-                </FormControl>
-              )}
-            </Stack>
-
-            {/* Optional Resource Overrides Section */}
-            <Divider />
-            <Stack spacing={1}>
-              <Stack
-                direction="row"
-                alignItems="center"
-                justifyContent="space-between"
-                sx={{ cursor: isSubmitting ? 'default' : 'pointer' }}
-                onClick={() => {
-                  if (!isSubmitting) {
-                    setShowResourceOverrides((prev) => !prev);
-                  }
-                }}
-              >
-                <Typography level="title-sm">
-                  Optional resource overrides
-                </Typography>
-                <ChevronDownIcon
-                  size={18}
-                  style={{
-                    transform: showResourceOverrides
-                      ? 'rotate(180deg)'
-                      : 'rotate(0deg)',
-                    transition: 'transform 0.2s ease',
-                  }}
-                />
-              </Stack>
-              {showResourceOverrides && (
-                <Stack spacing={2} ref={resourceOverridesRef}>
-                  <FormControl>
-                    <FormLabel>Override mode</FormLabel>
-                    <RadioGroup
-                      orientation="horizontal"
-                      value={resourceOverrideMode}
-                      onChange={(event) => {
-                        const nextMode = event.target.value as
-                          | 'manual'
-                          | 'group';
-                        setResourceOverrideMode(nextMode);
-                        if (nextMode === 'manual') {
-                          setSelectedResourceGroupId('');
-                          setResourceGroupCustomized(false);
-                        }
-                      }}
-                    >
-                      <Radio value="manual" label="Manual resources" />
-                      <Radio
-                        value="group"
-                        label="Use saved group"
-                        disabled={providerResourceGroups.length === 0}
-                      />
-                    </RadioGroup>
-                    {resourceOverrideMode === 'group' && (
-                      <FormHelperText>
-                        {providerResourceGroups.length === 0
-                          ? 'No saved groups are defined for this provider yet.'
-                          : 'Select a saved group to prefill resource values.'}
-                      </FormHelperText>
-                    )}
-                  </FormControl>
-
-                  {resourceOverrideMode === 'group' && (
-                    <FormControl>
-                      <FormLabel>Saved resource group</FormLabel>
-                      <Select
-                        placeholder="Select a resource group"
-                        value={selectedResourceGroupId || null}
-                        disabled={
-                          isSubmitting || providerResourceGroups.length === 0
-                        }
-                        onChange={(_, value) => {
-                          const groupId = String(value || '');
-                          setSelectedResourceGroupId(groupId);
-                          setResourceGroupCustomized(false);
-                          const group = providerResourceGroups.find(
-                            (entry) => entry.id === groupId,
-                          );
-                          if (!group) {
-                            return;
-                          }
-                          setCpusInput(group.cpus || '');
-                          setMemoryInput(group.memory || '');
-                          setDiskSpaceInput(group.disk_space || '');
-                          setAcceleratorsInput(group.accelerators || '');
-                          setNumNodesInput(group.num_nodes || '');
-                        }}
-                      >
-                        {providerResourceGroups.map((group) => (
-                          <Option key={group.id} value={group.id}>
-                            {group.name}
-                          </Option>
-                        ))}
-                      </Select>
-                      {selectedResourceGroupId && (
-                        <FormHelperText>
-                          {resourceGroupCustomized
-                            ? 'Custom override active: one or more fields have been edited.'
-                            : 'Using saved group values. Editing any field will switch to custom override.'}
-                        </FormHelperText>
-                      )}
-                    </FormControl>
-                  )}
-
-                  <Stack direction="row" spacing={2}>
-                    <FormControl sx={{ flex: 1 }}>
-                      <FormLabel>CPUs</FormLabel>
-                      <Input
-                        placeholder="e.g. 4"
-                        value={cpusInput}
-                        onChange={(e) => {
-                          setCpusInput(e.target.value);
-                          markCustomResourceOverride();
-                        }}
-                        disabled={isSubmitting}
-                      />
-                    </FormControl>
-                    <FormControl sx={{ flex: 1 }}>
-                      <FormLabel>Memory</FormLabel>
-                      <Input
-                        placeholder="e.g. 16GB"
-                        value={memoryInput}
-                        onChange={(e) => {
-                          setMemoryInput(e.target.value);
-                          markCustomResourceOverride();
-                        }}
-                        disabled={isSubmitting}
-                      />
-                    </FormControl>
-                  </Stack>
-                  <Stack direction="row" spacing={2}>
-                    <FormControl sx={{ flex: 1 }}>
-                      <FormLabel>Disk space</FormLabel>
-                      <Input
-                        placeholder="e.g. 100 or 100GB"
-                        value={diskSpaceInput}
-                        onChange={(e) => {
-                          setDiskSpaceInput(e.target.value);
-                          markCustomResourceOverride();
-                        }}
-                        disabled={isSubmitting}
-                      />
-                    </FormControl>
-                    <FormControl sx={{ flex: 1 }}>
-                      <FormLabel>Accelerators</FormLabel>
-                      <Input
-                        placeholder="e.g. A100:1, RTX3090:2, 1"
-                        value={acceleratorsInput}
-                        onChange={(e) => {
-                          setAcceleratorsInput(e.target.value);
-                          markCustomResourceOverride();
-                        }}
-                        disabled={isSubmitting}
-                      />
-                    </FormControl>
-                  </Stack>
-                  <Stack direction="row" spacing={2}>
-                    <FormControl sx={{ flex: 1 }}>
-                      <FormLabel>Num nodes</FormLabel>
-                      <Input
-                        placeholder="e.g. 1"
-                        value={numNodesInput}
-                        onChange={(e) => {
-                          setNumNodesInput(e.target.value);
-                          markCustomResourceOverride();
-                        }}
-                        disabled={isSubmitting}
-                      />
-                    </FormControl>
-                    <FormControl sx={{ flex: 1 }}>
-                      <FormLabel>Minutes requested</FormLabel>
-                      <Input
-                        type="number"
-                        placeholder="e.g. 60"
-                        value={minutesRequestedInput}
-                        onChange={(e) =>
-                          setMinutesRequestedInput(e.target.value)
-                        }
-                        disabled={isSubmitting}
-                      />
-                    </FormControl>
-                  </Stack>
-                  <FormHelperText>
-                    These values override the template&apos;s resource
-                    requirements for this run only. Leave a field empty to use
-                    the template default.
-                  </FormHelperText>
-
-                  {/* SkyPilot per-job overrides */}
-                  {isSkypilotProvider && (
-                    <>
-                      <Divider />
-                      <Typography level="title-sm">
-                        SkyPilot Job Overrides
-                      </Typography>
-                      <FormControl>
-                        <FormLabel>Docker Image (optional)</FormLabel>
-                        <Input
-                          value={jobDockerImage}
-                          onChange={(e) => setJobDockerImage(e.target.value)}
-                          placeholder="docker:nvcr.io/nvidia/pytorch:23.10-py3"
-                          sx={{ fontFamily: 'monospace', fontSize: 'sm' }}
-                          disabled={isSubmitting}
-                        />
-                        <FormHelperText>
-                          Prefix with &quot;docker:&quot; to run inside a
-                          container. Defaults to the provider&apos;s global
-                          setting.
-                        </FormHelperText>
-                      </FormControl>
-                      <FormControl>
-                        <FormLabel>Region (optional)</FormLabel>
-                        <Input
-                          value={jobRegion}
-                          onChange={(e) => setJobRegion(e.target.value)}
-                          placeholder="e.g. us-east-1"
-                          disabled={isSubmitting}
-                        />
-                      </FormControl>
-                      <FormControl
-                        sx={{ flexDirection: 'row', alignItems: 'center' }}
-                      >
-                        <Switch
-                          checked={jobUseSpot}
-                          onChange={(e) => setJobUseSpot(e.target.checked)}
-                          disabled={isSubmitting}
-                          sx={{ mr: 1 }}
-                        />
-                        <FormLabel sx={{ m: 0 }}>
-                          Use Spot / Preemptible Instances
-                        </FormLabel>
-                      </FormControl>
-                    </>
-                  )}
-
-                  {/* dstack per-job overrides */}
-                  {isDstackProvider && (
-                    <>
-                      <Divider />
-                      <Typography level="title-sm">
-                        dstack Job Overrides
-                      </Typography>
-                      <FormControl>
-                        <FormLabel>Fleet Name (optional)</FormLabel>
-                        <Input
-                          value={jobDstackFleetName}
-                          onChange={(e) =>
-                            setJobDstackFleetName(e.target.value)
-                          }
-                          placeholder="my-fleet"
-                          disabled={isSubmitting}
-                        />
-                        <FormHelperText>
-                          If set, this run is scheduled on the specified dstack
-                          fleet. Leave empty to use resource-based scheduling.
-                        </FormHelperText>
-                      </FormControl>
-                    </>
-                  )}
-
-                  {/* Incompatibility Warning */}
-                  {selectedProvider &&
-                    effectiveResources?.accelerators &&
-                    !isProviderCompatible(selectedProvider) && (
-                      <Alert
-                        variant="soft"
-                        color="warning"
-                        startDecorator={<AlertTriangleIcon size={18} />}
-                        sx={{ mt: 1 }}
-                      >
-                        <Typography level="body-sm">
-                          This provider may not support the requested
-                          accelerators (
-                          <strong>{effectiveResources.accelerators}</strong>).
-                        </Typography>
-                      </Alert>
-                    )}
-
-                  {/* Local Provider Resource Validation */}
-                  {isLocalProvider &&
-                    resourceValidation &&
-                    !resourceValidation.isCompatible && (
-                      <Alert
-                        variant="soft"
-                        color={
-                          resourceValidation.hasErrors ? 'danger' : 'warning'
-                        }
-                        startDecorator={<AlertTriangleIcon size={18} />}
-                        sx={{ mt: 1 }}
-                      >
-                        <Stack spacing={1}>
-                          <Typography
-                            level="title-sm"
-                            color={
-                              resourceValidation.hasErrors
-                                ? 'danger'
-                                : 'warning'
-                            }
-                          >
-                            {resourceValidation.hasErrors
-                              ? 'Local provider cannot meet task requirements'
-                              : 'Local provider may not meet task requirements'}
-                          </Typography>
-                          <Stack spacing={0.5}>
-                            {resourceValidation.issues.map((issue) => (
-                              <Stack
-                                key={`${issue.type}-${issue.label}`}
-                                direction="row"
-                                spacing={1}
-                                alignItems="center"
-                              >
-                                <Chip
-                                  size="sm"
-                                  variant="solid"
-                                  color={
-                                    issue.type === 'error'
-                                      ? 'danger'
-                                      : 'warning'
-                                  }
-                                >
-                                  {issue.label}
-                                </Chip>
-                                <Typography level="body-xs">
-                                  Required: <strong>{issue.required}</strong>,
-                                  Available: <strong>{issue.available}</strong>
-                                </Typography>
-                              </Stack>
-                            ))}
-                          </Stack>
-                          {resourceValidation.hasErrors && (
-                            <Typography level="body-xs" color="danger">
-                              Consider selecting a different provider with the
-                              required resources.
-                            </Typography>
-                          )}
-                        </Stack>
-                      </Alert>
-                    )}
-
-                  {isLocalProvider &&
-                    resourceValidation?.isCompatible &&
-                    effectiveResources && (
-                      <Alert
-                        variant="soft"
-                        color="success"
-                        startDecorator={<CheckCircleIcon size={18} />}
-                        sx={{ mt: 1 }}
-                      >
-                        <Typography level="body-sm" color="success">
-                          Local provider meets the task resource requirements.
-                        </Typography>
-                      </Alert>
-                    )}
-                </Stack>
-              )}
-            </Stack>
-          </Stack>
-        </DialogContent>
-        <DialogActions>
-          <Button variant="plain" color="neutral" onClick={onClose}>
-            Cancel
-          </Button>
-          <Button
-            startDecorator={<PlayIcon />}
-            color="success"
-            onClick={handleSubmit}
-            loading={isSubmitting}
-            disabled={!selectedProviderId || isSubmitting}
-          >
-            Queue task
-          </Button>
-          {isSubmitting && (
-            <Typography level="body-sm" sx={{ ml: 1 }}>
-              {loadingMessages[loadingMessageIndex]}
-            </Typography>
-          )}
-        </DialogActions>
-      </ModalDialog>
-    </Modal>
-  );
-}
